exercise-4
Exercise 4
----------
Randomly seed some cell with BOMB. Number of bombs is equal to board size (9, or 16, or 25)
Tapping on a BOMB cell ends the game, and displays a "BOOM" caption somewhere on screen
For an uncovered BOMB cell, display a Unicode Bomb character instead of "?". Or you may add an asset like a .png of a bomb; but in that case, check the asset into the project; do not reference outside resources.
Add a "Play Again" button to reset the game