exercise-5
Exercise 5
----------
Tapping any cell uncovers that cell AND ALSO its eight neighbors (fewer neighbors for cells on the edge of board, or if those neighbors are already uncovered)
The expansion is always just immediate-neighbor, no expansion beyond that.
Still displaying "?" or bomb in every uncovered cell.
Uncovering a neighboring bomb does not end the game. Only tapping directly on a bomb ends the game.
(None of this is standard Minesweeper rules, this is just for this exercise)