exercise-6
Exercise 6
----------
For cells that are uncovered, show the number of neighbors, covered or uncovered, that are bombs, instead of the "?". So each uncovered cell will show a number from 1 to 8, or a bomb.
(Note: Again this is not how Minesweeper works for real, it's only for this exercise.)
Keep in mind that we are still not building the "flood" expansion; still only uncovering one cell and up to eight neighboring cells at a time; but now uncovered cells will display numbers instead of "?"
Give some consideration to this question:
 - Should the numbers in each cell be pre-computed at the time that the board is generated, or should the numbers be computed when a cell is uncovered? Which of these options would be better if you were going to really build a "regulation" Minesweeper game?
In the code where this computation takes place, add inline documentation explaining your decision on which way to go with that.